update install/uninstall. comments

+++ install.sh
@@ -18,13 +18,13 @@ basedir=~/git/                                  # Where you want to put dotfiles
 dir=~/git/dotfiles                              # Your dotfiles directory
 budir=~/git/dotfiles_old                        # dotfiles_old directory
 
-# This section includes i3 config files. Comment out these two lines and the i3 section 
-# further down if you aren't using i3
-files="i3status.conf vimrc Xresources zshrc"    # Files to symlink
-i3conf=i3config                                 # ~/.config/i3/config name in backup/dotfiles
-
-# Uncomment the line below if don't have i3
-#files="vimrc Xresources zshrc"    # Files to symlink
+if [ -f "$HOME/.i3status.conf" ] && [ -f "$HOME/.config/i3/config" ] ; then
+  echo "i3 Detected..."
+  files="i3status.conf vimrc Xresources zshrc"    # Files to symlink
+  i3conf=i3config                                 # ~/.config/i3/config name in backup/dotfiles
+else
+  files="vimrc Xresources zshrc"                  # Files to symlink
+fi
 
 #---------------------------------------------------------------------------------------------
 # End Variables
@@ -49,6 +49,7 @@ echo "...done"
 
 # Copy all the existing dotfiles into the backup dir, 
 # and create symlinks to dotfiles in your git dotfiles folder. 
+# This section takes care of any dotfiles located in your Homedir
 
 for file in $files; do
   echo "Backing up $file to $budir"
@@ -61,20 +62,21 @@ done
 
 
 #---------------------------------------------------------------------------------------------
-# Program ends here if you don't have i3. comment out everything below
+# Program ends here if you don't have i3.
 #---------------------------------------------------------------------------------------------
 
-# i3 section. Uncomment these following lines if you aren't using i3!!
-# Because the i3config file is in a different place I put it here seperately.
-# This just does the same thing, backs up the existing file to dotfiles_old,
+# Because the i3config file is in a different place I put this here seperately.
+# It just does the same thing, backs up the existing file to dotfiles_old,
 # and then creates a symlink from dotfiles/i3config. 
 
-echo "Backing up $i3conf to $budir"
-mv ~/.config/i3/config $budir/i3config
-echo "...done"
-echo "Creating symlink to i3config in ~"
-ln -s $dir/$i3conf ~/.config/i3/config
-echo "...done"
+if [ -f "$HOME/.i3status.conf" ] && [ -f "$HOME/.config/i3/config" ] ; then
+  echo "Backing up $i3conf to $budir"
+  mv ~/.config/i3/config $budir/i3config
+  echo "...done"
+  echo "Creating symlink to i3config in ~"
+  ln -s $dir/$i3conf ~/.config/i3/config
+  echo "...done"
+fi
 
 #---------------------------------------------------------------------------------------------
 # Program End
